+++ R/rcax_citation.R
@@ -2,9 +2,9 @@
 #'
 #' @export
 #' @return API citation as character string
-#' @examples \dontrun{
+#' @examples
 #' rcax_citation()
-#' }
+#' 
 rcax_citation <- function(...) {
-  sprintf('Coordinated Assessments. Version %s <https://www.streamnet.org/home/data-maps/fish-hlis/>')
+  sprintf('Coordinated Assessments. <https://www.streamnet.org/home/data-maps/fish-hlis/>')
 }