+++ R/rcax_escdata.R
@@ -8,57 +8,30 @@
 #'  `rcax_escdata()` will download 1000 records. 
 #'  
 #' @export
-#' @template all
-#' @template commonargs
+#' @template tableargs
 #' @template info
 #' @examples 
-#' # to get all columns and print the column names
-#' # a <- rcax_escdata(cols = NULL)
-#' # colnames(a)
+#' rcax_escdata(cols = NULL, qlist=list(limit=5))
 #' 
-#' @references 
-#' This function is modeled off the functions in \url{https://github.com/ropensci/rredlist}
-#' 
-rcax_escdata <- function(
-    extra = NULL, 
-    table_id = NULL,
-    key = NULL, parse = TRUE, 
-    cols = NULL, 
-    type = c("data.frame", "colnames"), 
-    tablename = "EscData", 
-    sortcols = c("countdate", "refid"), ...) {
-  # error checking
-  type <- match.arg(type)
-  assert_is(table_id, 'character')
-  assert_is(extra, 'list')
-  assert_is(parse, 'logical')
-  
-  # get the table_id for the table
-  if(is.null(table_id)) table_id = subset(rCAX:::caxtabs, name==tablename)$id
-  
-  # set up query list
-  query_list <- list(table_id = table_id)
-  if(!is.null(extra)) query_list <- c(query_list, extra)
-  # if just returning colnames, set limit to 1
-  if(type=="colnames") query_list$limit <- 1
-  
-  # Make API call
-  tab <- rcax_parse(rcax_GET("ca", key, query=query_list, ...), parse)
-  tab <- tab$records
-  
-  # if type is colnames, return that
-  if(type=="colnames") return(colnames(tab))
-  
-  # if type is data.frame, filter and sort
-  for(i in sortcols)
-    if(sortcols[i] %in% colnames(tab)) tab <- tab[order(tab[sortcols[i]]),]
-  if(!is.null(cols)) tab <- tab[,cols]
-  tab
-}
+#' # to print the first 5 column names
+#' rcax_escdata(type="colnames")[1:5]
 
-# Not used; this would return just the json
-rcax_escdata_ <- function(table_id, extra = NULL, key = NULL, ...) {
-  assert_is(table_id, 'character')
-  assert_is(key, 'character')
-  rcax_GET("ca", key, query=c(list(table_id=table_id), extra), ...)
+rcax_escdata <- function(
+    tablename = "EscData", 
+    flist = NULL,
+    qlist = NULL, 
+    cols = NULL, 
+    sortcols = c("countdate", "refid"),
+    type = c("data.frame", "colnames"), 
+    GETargs = list(table_id = NULL, recordloc = "records", key = NULL, parse = TRUE),  ...) {
+  
+  # API call and table filtering and sorting
+  rcax_table_query(
+    tablename = tablename, 
+    flist = flist, qlist = qlist, 
+    cols = cols, sortcols = sortcols,
+    type = type, 
+    GETargs = list(recordloc = "records"), ...)
+  
+  
 }